mcd_dart/requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
imagehash>=4.2.0
scipy>=1.6.0
shapely>=1.7.0
matplotlib>=3.3.0